+++ client/src/pages/homepage/homePage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import '../styles/HomePage.css'; // Create a new CSS file for homepage styles
+import '../../styles/HomePage.css'; // Create a new CSS file for homepage styles
 
 const HomePage = () => {
   const theme = useSelector(state => state.theme); // Assuming you have a theme state in Redux
